// Work/WK/MS/Web_WebSV/Web/Web.Core/wwwroot/CTS/ClientApp/yinsiang-lottery-fullscreen/src/models/LotteryMenuInfo.ts
export interface LotteryMenuInfo {
  gameTypeID: number;
  groupPriority: number;
  hotNew: number;
  lotteryID: number;
  lotteryType: string;
  maxBonusMoney: number;
  notice: string;
  numberTrendUrl: string;
  officialLotteryUrl: string;
  priority: number;
  typeURL: string;
  userType: number;
  isMaintaining: boolean;
}